requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg[binary]==3.2.9
